Added Access Control List

+++ app/Http/Controllers/Admin/FooController.php
@@ -7,6 +7,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Redirect;
+use Illuminate\Support\Facades\Gate;
 
 use App\Foo;
 
@@ -26,6 +27,9 @@ class FooController extends Controller
      */
     public function index(Request $request)
     {
+        if(Gate::denies('foos-view'))
+            return Redirect::back()->withErrors([trans('crud.not-authorized')]);
+            
         $query = $request->get('q');
         $items = Foo::where('something', 'ilike', "%{$query}%")
                     ->where('user_id', Auth::user()->id)
@@ -40,6 +44,9 @@ class FooController extends Controller
      */
     public function create()
     {
+        if(Gate::denies('foos-create'))
+            return Redirect::back()->withErrors([trans('crud.not-authorized')]);
+
         return view('admin.foo.form');
     }
 
@@ -51,6 +58,9 @@ class FooController extends Controller
      */
     public function store(Request $request)
     {
+        if(Gate::denies('foos-create'))
+            return Redirect::back()->withErrors([trans('crud.not-authorized')]);
+
         $fields = $request->item;
         $validator = Validator::make($fields, $this->rules, [], $this->names);
         if ($validator->fails())
@@ -81,10 +91,16 @@ class FooController extends Controller
      */
     public function show($id)
     {
+        if(Gate::denies('foos-view'))
+            return Redirect::back()->withErrors([trans('crud.not-authorized')]);
+
         $item = Foo::find($id);
         if(!isset($item))
             return Redirect::back()->withErrors([trans('crud.item-not-found')]);
             
+        if(Gate::denies('mine', $item))
+            return Redirect::back()->withErrors([trans('crud.not-authorized')]);
+        
         return view('admin.foo.show', compact('item'));
     }
 
@@ -96,10 +112,16 @@ class FooController extends Controller
      */
     public function edit($id)
     {
+        if(Gate::denies('foos-update'))
+            return Redirect::back()->withErrors([trans('crud.not-authorized')]);
+
         $item = Foo::find($id);
         if(!isset($item))
             return Redirect::back()->withErrors([trans('crud.item-not-found')]);
 
+        if(Gate::denies('mine', $item))
+            return Redirect::back()->withErrors([trans('crud.not-authorized')]);
+        
         return view('admin.foo.form', compact('item'));
     }
 
@@ -112,6 +134,14 @@ class FooController extends Controller
      */
     public function update(Request $request, $id)
     {
+        if(Gate::denies('foos-update'))
+            return Redirect::back()->withErrors([trans('crud.not-authorized')]);
+
+        $item = Foo::find($id);
+            
+        if(Gate::denies('mine', $item))
+            return Redirect::back()->withErrors([trans('crud.not-authorized')]);
+
         $fields = $request->item;
         $validator = Validator::make($fields, $this->rules, [], $this->names);
         if ($validator->fails())
@@ -122,7 +152,6 @@ class FooController extends Controller
         }
         
         try {
-            $item = Foo::find($id);
             $item->something = $fields['something'];
             $item->save();
 
@@ -143,6 +172,9 @@ class FooController extends Controller
      */
     public function destroy($id)
     {
+        if(Gate::denies('foos-delete'))
+            return Redirect::back()->withErrors([trans('crud.not-authorized')]);
+
         try {
             $item = Foo::find($id);
             $item->delete();
